requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.9
confusable-homoglyphs==3.2.0
cryptography==3.4.6
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filer==2.0.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-registration==3.1.1
django-resized==0.3.11
django-rest-auth==0.9.5
django-vite==1.2.1
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.4.0
easy-thumbnails==2.7.1
idna==2.10
lxml==4.7.1
mysql-connector-python==8.0.23
oauthlib==3.1.0
Pillow==8.4.0
protobuf==3.15.6
pycparser==2.20
PyJWT==2.0.1
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.1
reportlab==3.6.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
svglib==1.1.0
tinycss2==1.1.1
Unidecode==1.1.2
urllib3==1.26.4
webencodings==0.5.1

opencv-python~=4.5.5.62
pytesseract==0.3.7
mojimoji~=0.0.12